+++ R/create_informant.R
@@ -11,7 +11,7 @@
 #  
 #  This file is part of the 'rstudio/pointblank' project.
 #  
-#  Copyright (c) 2017-2023 pointblank authors
+#  Copyright (c) 2017-2024 pointblank authors
 #  
 #  For full copyright and license information, please look at
 #  https://rstudio.github.io/pointblank/LICENSE.html
@@ -55,15 +55,15 @@
 #' 
 #' @param tbl *Table or expression for reading in one*
 #' 
-#'   `obj:<tbl_*>|<table reading expression>` // **required**
+#'   `obj:<tbl_*>|<tbl reading expression>` // **required**
 #' 
 #'   The input table. This can be a data frame, a tibble, a `tbl_dbi` object, or
 #'   a `tbl_spark` object. Alternatively, an expression can be supplied to serve
 #'   as instructions on how to retrieve the target table at incorporation-time.
 #'   There are two ways to specify an association to a target table: (1) as a
 #'   table-prep formula, which is a right-hand side (RHS) formula expression
-#'   (e.g., `~ { <table reading code>}`), or (2) as a function (e.g.,
-#'   `function() { <table reading code>}`).
+#'   (e.g., `~ { <tbl reading code>}`), or (2) as a function (e.g.,
+#'   `function() { <tbl reading code>}`).
 #'   
 #' @param agent *The pointblank agent object*
 #' 
@@ -399,6 +399,10 @@ create_informant <- function(
     }
   }
   
+  private <- list(
+    col_ptypes = tbl_info$col_ptypes
+  )
+  
   metadata_list <-
     c(
       list(
@@ -409,7 +413,8 @@ create_informant <- function(
           `_type` = table_type
         )
       ),
-      column_list
+      column_list,
+      list(`_private` = private)
     )
 
   # Create the metadata list object